feat: follow up task added a better soliution
